增加支持SDM15单点雷达

+++ src/CYdLidar.h
@@ -1110,7 +1110,7 @@ class YDLIDAR_API CYdLidar {
   int m_SerialBaudrate;             ///< LiDAR serial baudrate or network port
   int m_LidarType;                  ///< LiDAR type
   int m_DeviceType;                 ///< LiDAR device type
-  int m_SampleRate;                 ///< LiDAR sample rate
+  float m_SampleRate;                 ///< LiDAR sample rate
   int m_SampleRatebyD1;             ///< LiDAR sample rate by d1
   int m_AbnormalCheckCount;         ///< LiDAR abnormal count
